+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+## Unreleased
+### Fixed
+- Fixed protocol loss during redis cluster `MOVED` / `ASK`
+
 ## v2.4.0-RC1 (2024-11-21)
 ### Added
 - Added new hash-field expiration commands (#1520)